Massive refactor to try to organize this so that it's managable
Additionally made responses and parameters as separate as possible so to favor decoupling the different sports from each other. Possibly at the end there will be another effort to refactor and reuse duplicated components but in general it'll be less of a headache to just create a parameter class for each request and a response payload for each request.